major code cleanup: removed IResourceManager and ICommunicator
git-svn-id: http://morgoth.zemris.fer.hr/svn/vhdllab/vhdllab2/trunk@1037 cbae1e92-611c-0410-ab7b-b19ac7dee622

+++ vhdllab-client/src/main/java/hr/fer/zemris/vhdllab/platform/manager/file/DefaultFileManager.java
@@ -2,6 +2,7 @@ package hr.fer.zemris.vhdllab.platform.manager.file;
 
 import hr.fer.zemris.vhdllab.api.workspace.FileReport;
 import hr.fer.zemris.vhdllab.entities.FileInfo;
+import hr.fer.zemris.vhdllab.entities.FileType;
 import hr.fer.zemris.vhdllab.platform.listener.AbstractEventPublisher;
 import hr.fer.zemris.vhdllab.platform.manager.workspace.WorkspaceManager;
 import hr.fer.zemris.vhdllab.service.FileService;
@@ -43,8 +44,10 @@ public class DefaultFileManager extends AbstractEventPublisher<FileListener>
     @Override
     public void delete(FileInfo file) {
         checkIfNull(file);
-        FileReport report = service.delete(file);
-        fireFileDeleted(report);
+        if(!file.getType().equals(FileType.PREDEFINED)) {
+            FileReport report = service.delete(file);
+            fireFileDeleted(report);
+        }
     }
 
     private void fireFileCreated(FileReport report) {
